+++ src/hybridmap/HybridNeutralAtomMapper.cpp
@@ -11,6 +11,7 @@
 #include "hybridmap/HybridNeutralAtomMapper.hpp"
 
 #include "circuit_optimizer/CircuitOptimizer.hpp"
+#include "datastructures/CircuitOptimizations.hpp"
 #include "hybridmap/Mapping.hpp"
 #include "hybridmap/MoveToAodConverter.hpp"
 #include "hybridmap/NeutralAtomDefinitions.hpp"
@@ -68,12 +69,12 @@ void NeutralAtomMapper::mapAppend(qc::QuantumComputation& qc,
     mappedQc.addAncillaryRegister(this->arch->getNpositions(), "fa");
   }
 
-  qc::CircuitOptimizer::replaceMCXWithMCZ(qc);
+  qmap::replaceMCXWithMCZ(qc);
   qc::CircuitOptimizer::singleQubitGateFusion(qc);
   qc::CircuitOptimizer::flattenOperations(qc);
   qc::CircuitOptimizer::removeFinalMeasurements(qc);
 
-  const auto dag = qc::CircuitOptimizer::constructDAG(qc);
+  const auto dag = constructDAG(qc);
 
   mapping = initialMapping;
 
@@ -168,10 +169,10 @@ void NeutralAtomMapper::decomposeBridgeGates(qc::QuantumComputation& qc) const {
 
 qc::QuantumComputation NeutralAtomMapper::convertToAod() {
   // decompose SWAP gates
-  qc::CircuitOptimizer::decomposeSWAP(mappedQc, false);
+  qmap::decomposeSWAP(mappedQc, false);
   // decompose bridge gates
   decomposeBridgeGates(mappedQc);
-  qc::CircuitOptimizer::replaceMCXWithMCZ(mappedQc);
+  qmap::replaceMCXWithMCZ(mappedQc);
   qc::CircuitOptimizer::singleQubitGateFusion(mappedQc);
   qc::CircuitOptimizer::flattenOperations(mappedQc);
   // decompose AOD moves