+++ src/hybridmap/NeutralAtomUtils.cpp
@@ -11,6 +11,7 @@
 #include "hybridmap/NeutralAtomUtils.hpp"
 
 #include "circuit_optimizer/CircuitOptimizer.hpp"
+#include "datastructures/CircuitOptimizations.hpp"
 #include "ir/Definitions.hpp"
 #include "ir/QuantumComputation.hpp"
 #include "ir/operations/OpType.hpp"
@@ -142,7 +143,7 @@ void BridgeCircuits::computeBridgeCircuit(const size_t length) {
   qcBridge.insert(qcBridge.begin(), std::make_unique<qc::StandardOperation>(
                                         qcBridge.getNqubits() - 1, qc::H));
 
-  qc::CircuitOptimizer::replaceMCXWithMCZ(qcBridge);
+  qmap::replaceMCXWithMCZ(qcBridge);
   qc::CircuitOptimizer::singleQubitGateFusion(qcBridge);
   bridgeCircuits[length] = qcBridge;
 }